+++ realm/realm-library/src/main/java/io/realm/internal/UncheckedRow.java
@@ -188,6 +188,13 @@ public class UncheckedRow implements NativeObject, Row {
         return new LinkView(context, parent, columnIndex, nativeLinkViewPtr);
     }
 
+    @Override
+    public void checkIfBacklinkAvailable() {
+        if (!isAttached()) {
+            throw new IllegalStateException("Object is no longer managed by Realm. Has it been deleted?");
+        }
+    }
+
     // Setters
 
     @Override